Moved all the different functions to their own cpp
