+++ Crystal-Home-Systems/Crystal-Home-Systems.vcxproj
@@ -75,9 +75,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Command.cpp" />
     <ClCompile Include="HomeSystem.cpp" />
+    <ClCompile Include="Media.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="command.h" />
+    <ClInclude Include="homeSystem.h" />
+    <ClInclude Include="media.h" />
     <ClInclude Include="sysParams.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
