+++ Crystal-Home-Systems/Debug/Crystal-Home-Systems.log
@@ -1,14 +1,18 @@
-﻿Build started 5/20/2015 9:03:34 PM.
+﻿Build started 5/20/2015 9:24:08 PM.
      1>Project "C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Crystal-Home-Systems\Crystal-Home-Systems.vcxproj" on node 2 (Build target(s)).
      1>ClCompile:
          C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\CL.exe /c /I"C:\SFML-2.3\include" /ZI /nologo /W3 /WX- /sdl /Od /Oy- /D _MBCS /Gm /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Fo"Debug\\" /Fd"Debug\vc120.pdb" /Gd /TP /analyze- /errorReport:prompt HomeSystem.cpp
          HomeSystem.cpp
      1>c:\users\austin\documents\visual studio 2013\projects\crystal-home-systems\crystal-home-systems\sysparams.h(28): warning C4305: 'initializing' : truncation from 'double' to 'float'
        Link:
-         C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\link.exe /ERRORREPORT:PROMPT /OUT:"C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Debug\Crystal-Home-Systems.exe" /INCREMENTAL /NOLOGO /LIBPATH:"C:\SFML-2.3\lib" "sfml-audio-d.lib" "sfml-graphics-d.lib" "sfml-window-d.lib" "sfml-network-d.lib" "sfml-system-d.lib" kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /MANIFEST /MANIFESTUAC:"level='asInvoker' uiAccess='false'" /manifest:embed /DEBUG /PDB:"C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Debug\Crystal-Home-Systems.pdb" /TLBID:1 /DYNAMICBASE /NXCOMPAT /IMPLIB:"C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Debug\Crystal-Home-Systems.lib" /MACHINE:X86 Debug\HomeSystem.obj
-         Crystal-Home-Systems.vcxproj -> C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Debug\Crystal-Home-Systems.exe
-     1>Done Building Project "C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Crystal-Home-Systems\Crystal-Home-Systems.vcxproj" (Build target(s)).
+         C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\link.exe /ERRORREPORT:PROMPT /OUT:"C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Debug\Crystal-Home-Systems.exe" /INCREMENTAL /NOLOGO /LIBPATH:"C:\SFML-2.3\lib" "sfml-audio-d.lib" "sfml-graphics-d.lib" "sfml-window-d.lib" "sfml-network-d.lib" "sfml-system-d.lib" kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /MANIFEST /MANIFESTUAC:"level='asInvoker' uiAccess='false'" /manifest:embed /DEBUG /PDB:"C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Debug\Crystal-Home-Systems.pdb" /TLBID:1 /DYNAMICBASE /NXCOMPAT /IMPLIB:"C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Debug\Crystal-Home-Systems.lib" /MACHINE:X86 Debug\Command.obj
+         Debug\HomeSystem.obj
+         Debug\Media.obj
+     1>HomeSystem.obj : error LNK2005: "class sf::Music music" (?music@@3VMusic@sf@@A) already defined in Command.obj
+     1>Media.obj : error LNK2005: "class sf::Music music" (?music@@3VMusic@sf@@A) already defined in Command.obj
+     1>C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Debug\Crystal-Home-Systems.exe : fatal error LNK1169: one or more multiply defined symbols found
+     1>Done Building Project "C:\Users\Austin\documents\visual studio 2013\Projects\Crystal-Home-Systems\Crystal-Home-Systems\Crystal-Home-Systems.vcxproj" (Build target(s)) -- FAILED.
 
-Build succeeded.
+Build FAILED.
 
-Time Elapsed 00:00:01.35
+Time Elapsed 00:00:01.44
