+++ Crystal-Home-Systems/HomeSystem.cpp
@@ -25,28 +25,18 @@
 */
 
 #include "sysParams.h"
-#include <stdlib.h>
+#include "media.h"
+#include "command.h"
+#include "homeSystem.h"
 #include <iostream>
-#include <algorithm>
 #include <string>
 #include <windows.h>
 
-#include <SFML/Audio.hpp>
-
 using namespace std;
-using namespace sf;
-
-void fileOpenError();
-string getCommand(char[]);
-void systemStartupMessage();
-void playMusic(char[]);
-void analyzeCommand(string, bool);
-
-Music music; // needs to be global for the playmusic function, cannot share var through passing
 
 int main()
 {
-	bool running = true; 
+	bool running = true;
 	string command;
 
 	// System Setup and Title
@@ -68,98 +58,7 @@ void fileOpenError(){
 	cout << "Unable to open file!" << endl;
 }
 
-string getCommand(char c[]){
-	cout << c << endl;
-	string command;
-	getline(cin, command);
-	transform(command.begin(), command.end(), command.begin(), ::tolower);
-	return command;
-}
-
 void systemStartupMessage(){
 	SetConsoleTitle(TEXT("Crystal Home Systems")); // set console window title to Crystal Home System
 	cout << systemName << " " << systemType << " Version " << version << "\nCreated By Ezra and Austin\n\n" << endl;
-}
-
-void playMusic(char dir[]){
-	if (!music.openFromFile(dir))
-		fileOpenError();
-	music.play();
-}
-
-void analyzeCommand(string command, bool running){
-	// finding command terms in the string. to add more commands to an if statement, enter: command.find("putCommandYouWandToUseHere") != string::npos
-	if (command.find("exit") != string::npos || command.find("quit") != string::npos || command.find("close") != string::npos){ // user wants to quit
-		if (command.find("song") != string::npos || command.find("music") != string::npos){
-			music.stop();
-			return;
-		}
-		if (command.find("movie") != string::npos || command.find("show") != string::npos){
-			return;
-		}
-		running = false;
-		music.stop();
-		exit(0);
-	}
-
-	// stop command
-	if (command.find("stop") != string::npos){
-		if (command.find("song") != string::npos || command.find("music") != string::npos){
-			music.stop();
-			return;
-		}
-		if (command.find("movie") != string::npos || command.find("show") != string::npos){
-			return;
-		}
-	}
-
-	// play command
-	if (command.find("play") != string::npos || command.find("start") != string::npos){ // wants to play media
-		if (command.find("movie") != string::npos || command.find("show") != string::npos){ // movie
-
-		}
-		else if (command.find("song") != string::npos || command.find("music") != string::npos || command.find("album") != string::npos || command.find("artist") != string::npos){ // music
-			// now that the loop as narrowed down that you want to play a song, make sure there's not another song already loaded, and then start the new song
-			if (music.Paused){
-				cout << "There is already a song playing. Are you sure you want to start a new one?" << endl;
-				string temp;
-				getline(cin, temp);
-				transform(temp.begin(), temp.end(), temp.begin(), ::tolower);
-				if (temp.find("no") != string::npos){ // dont play new song
-					music.play();
-					return;
-				}
-				else if (temp.find("yes") != string::npos){
-					// LOAD SONG HERE
-					return;
-				}
-				cout << "Please enter yes or no!" << endl;
-				return;
-			}
-		}
-	}
-
-	// pause command
-	if (command.find("pause") != string::npos){
-		if (command.find("music") != string::npos || command.find("song") != string::npos){
-			music.pause();
-		}
-		if (command.find("movie") != string::npos || command.find("show") != string::npos || command.find("tv") != string::npos){
-
-		}
-	}
-
-	// resume command
-	if (command.find("resume") != string::npos){
-		if (command.find("music") != string::npos || command.find("song") != string::npos){
-			if (music.Paused){
-				music.play();
-				return;
-			}
-			else{
-				cout << "There is no song to resume!" << endl;
-				return;
-			}
-		}
-	}
 }
