+++ Crystal-Home-Systems/command.h
@@ -16,17 +16,7 @@
 *along with Crystal Home Systems.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-/*
-Global System Parameters
-*/
+#include <string>
 
-char systemName[] = "Crystal Home Systems";
-
-//char systemType[] = "Shard";
-char systemType[] = "Server";
-
-float version = 0.1;
-
-char musicPath[] = "";
-
-char moviePath[] = "";
+void analyzeCommand(std::string, bool);
+std::string getCommand(char[]);
